Add files via upload
Just a basic batch password generator. It will not repeat characters, and excludes a few commonly rejected characters like parentheses, and tilde, etc. Feel free to add your own. I had not practiced coding in a while, so this is one of the first things I have tried in my journey back to learning.